# submission/03-文档文件夹/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 区块链电子病历共享系统API
  version: 1.0.0
  description: |
    基于区块链的电子病历共享系统RESTful API文档

    ## 功能特性
    - 🔐 基于JWT的用户认证
    - 📁 病历文件上传和管理
    - 🔒 细粒度权限控制
    - 🔗 区块链存证和追溯
    - 📊 完整的审计日志

    ## 认证方式
    所有API都需要在请求头中包含有效的JWT令牌：
    ```
    Authorization: Bearer <your-jwt-token>
    ```
  contact:
    name: EMR System Support
    email: support@emr-system.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 开发环境
  - url: https://api.emr-system.com/v1
    description: 生产环境

# 安全模式定义
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT令牌认证
  schemas:
...
    MedicalRecord:
      type: object
      properties:
        recordId:
          type: string
          format: uuid
          description: 病历记录ID
        patientId:
          type: string
          format: uuid
          description: 患者ID
        creatorId:
          type: string
          format: uuid
          description: 创建者ID
        title:
          type: string
          description: 病历标题
        description:
          type: string
          description: 病历描述
        recordType:
          type: string
          enum: [CT, MRI, X_RAY, ECG, BLOOD_TEST, PATHOLOGY, OTHER]
          description: 病历类型
        department:
          type: string
          description: 科室
        ipfsCid:
          type: string
          description: IPFS内容标识符
        blockchainTxId:
          type: string
          description: 区块链交易ID
        fileSize:
          type: integer
          description: 文件大小（字节）
        fileHash:
          type: string
          description: 文件SHA-256哈希
        createdAt:
          type: string
          format: date-time
          description: 创建时间
...
security:
  - bearerAuth: []
paths:
...
  /records:
    get:
      tags:
        - Medical Records
      summary: 获取病历列表
      description: 获取当前用户可访问的病历列表
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [created_at, title, record_type]
            default: created_at
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: recordType
          in: query
          schema:
            type: string
            enum: [CT, MRI, X_RAY, ECG, BLOOD_TEST, PATHOLOGY, OTHER]
        - name: department
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
          description: 搜索关键词
      responses:
        "200":
          description: 成功获取病历列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    type: array
                    items:
                      $ref: "#/components/schemas/MedicalRecord"
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer

    post:
      tags:
        - Medical Records
      summary: 上传新病历
      description: 上传新的病历文件到系统
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, patientId, title, recordType]
              properties:
                file:
                  type: string
                  format: binary
                  description: 病历文件
                patientId:
                  type: string
                  format: uuid
                  description: 患者ID
                title:
                  type: string
                  description: 病历标题
                description:
                  type: string
                  description: 病历描述
                recordType:
                  type: string
                  enum: [CT, MRI, X_RAY, ECG, BLOOD_TEST, PATHOLOGY, OTHER]
                department:
                  type: string
                  description: 科室
      responses:
        "201":
          description: 病历上传成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  recordId:
                    type: string
                    format: uuid
                  ipfsCid:
                    type: string
                  blockchainTxId:
                    type: string
                  message:
                    type: string
  /records/{recordId}:
    get:
      tags:
        - Medical Records
      summary: 获取病历详情
      description: 获取指定病历的详细信息
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 成功获取病历详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/MedicalRecord"
                  - type: object
                    properties:
                      content:
                        type: string
                        description: 解密后的病历内容（Base64编码）
        "403":
          description: 无权限访问该病历
        "404":
          description: 病历不存在
...